% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A function that uses the WorldPopulation dataframe
to generate a graph of any country in the WorldPopulation dataframe.}
\usage{
CountryPopulation(country_name, title = NULL)
}
\arguments{
\item{title}{title for the garph}

\item{Country_name}{Name of the country}
}
\value{
return the graph of the country
}
\description{
A function that uses the WorldPopulation dataframe
to generate a graph of any country in the WorldPopulation dataframe.
}
\examples{
CountryPopulation("Brazil")
CountryPopulation("Brazil",'Graph of Brazil population over time')
}
