% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup_doc.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{FIFA World Cup Statistics (1930-2022).}
\format{
a data frame with 22 different FIFA World Cup tournaments.
\describe{
  \item{Year}{The Year at which the FIFA World Cup was played.}
  \item{Hosts}{The Country that hosted the FIFA World Cup.}
  \item{Matches}{Total number of matches played.}
  \item{Totalattendance}{Total attendance across all matches in the tournament.}
  \item{Averageattendance}{Average attendance per match in the tournament.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
FIFA World Cup Statistics (1930-2022).
}
\keyword{datasets}
