% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Population data for UN member countries from 1950 to 2020.}
\format{
a data frame with 235 observations
\describe{
  \item{Country_Name}{Names of all (235) countries that are part of the UN.}
  \item{Year=`1950` to `2020`}{Year at which the population was taken.}
}
}
\source{
\url{https://population.un.org/wpp/}
}
\usage{
WorldPopulation
}
\description{
Population data for UN member countries from 1950 to 2020.
}
\keyword{datasets}
